ci: unify github actions for merge and pull request
There's no need to do anything different and last merge showed it's not good when the
actions drift apart. PR coverage was ok, but merge coverage is abysmal since the action
fell behind.

And while we're at it, remove the `// ignore.coverage` comment from lines where it has no
effect.